% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_recursive.R
\name{get_data_recursive}
\alias{get_data_recursive}
\title{wratpper for get_climate_data()}
\usage{
get_data_recursive(locator, geo_type, type, cvar, start, end)
}
\arguments{
\item{locator}{The ISO3 country code that you want data about. (http://unstats.un.org/unsd/methods/m49/m49alpha.htm) or the basin ID [1-468]}

\item{geo_type}{basin or country depending on the locator type}

\item{type}{the type of data you want "mavg" for monthly averages, "annualavg"}

\item{cvar}{The variable you're interested in. "pr" for precipitation, "tas" for temperature in celcius.}

\item{start}{The starting year you want data for, can be in the past or the future. Must conform to the periods outlined in the world bank API.  If given values don't conform to dates, the fuction will automatically round them.}

\item{end}{The ending year you want data for, can be in the past or the future.  Similar to the start date, dates will be rounded to the nearest end dat.}
}
\description{
Function to recursively call the get_climate_data().  Handles a vector of basins
or countries as well as multiple dates.
}
\examples{
\dontrun{
 get_ensemble_data_recursive(c("1","2"),"basin","mavg","pr",1920,1940)
}
}
